TFC_PB/requirements.txt
django==4.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
geopy==2.2.0
pillow==9.2.0
python-dateutil==2.8.2
django-cors-headers==3.13.0